Create onboardAuth2Tokens.py
When onboarding participants on the first day of the study, after having then navigate through the oath url to consent to allowing the stanford research app to access their data, use this code to retrieve the access and refresh tokens from their auth codes. You can manually add their tokens to the firestore backend or add some could to upload it to firestore automatically.